Move configuration and constants out of models
