+++ mchub/models/cloud/openstack_manager.py
@@ -3,7 +3,7 @@ import openstack
 from os import environ, path
 from re import search, IGNORECASE
 
-from .. constants import INSTANCE_CATEGORIES
+from ... constants import INSTANCE_CATEGORIES
 
 CENTOS_VALID_IMAGES = r"centos"
 OTHER_VALID_IMAGES = r"almalinux|rocky"
