+++ mchub/resources/api_view.py
@@ -4,10 +4,10 @@ from flask import request
 from flask.views import MethodView
 from flask import make_response
 
+from .. configuration import config
 from .. database.database_manager import DatabaseManager
 from .. models.auth_type import AuthType
 from .. models.user.anonymous_user import AnonymousUser
-from .. models.configuration import config
 from .. models.user.authenticated_user import AuthenticatedUser
 from .. exceptions.invalid_usage_exception import (
     UnauthenticatedException,
