+++ tests/integration/test_api_no_auth.py
@@ -1,5 +1,5 @@
+from mchub.constants import DEFAULT_CLOUD
 from mchub.models.magic_castle.cluster_status_code import ClusterStatusCode
-from mchub.models.constants import DEFAULT_CLOUD
 from mchub import app
 
 from .. test_helpers import *
