+++ tests/test_helpers.py
@@ -7,10 +7,10 @@ from shutil import rmtree, copytree
 from typing import Callable
 
 
+from mchub.constants import DEFAULT_CLOUD
 from mchub.database.schema_manager import SchemaManager
 from mchub.database.database_manager import DatabaseManager
 from mchub.models.user.authenticated_user import AuthenticatedUser
-from mchub.models.constants import DEFAULT_CLOUD
 
 from . mocks.openstack.openstack_connection_mock import OpenStackConnectionMock
 
@@ -176,7 +176,7 @@ def admin() -> Callable[[sqlite3.Connection], AuthenticatedUser]:
 
 @pytest.fixture(autouse=True)
 def mock_clusters_path(mocker):
-    mocker.patch("mchub.models.constants.CLUSTERS_PATH", new=MOCK_CLUSTERS_PATH)
+    mocker.patch("mchub.constants.CLUSTERS_PATH", new=MOCK_CLUSTERS_PATH)
     mocker.patch(
         "mchub.models.magic_castle.magic_castle.CLUSTERS_PATH",
         new=MOCK_CLUSTERS_PATH,
